add list of care charge package
