+++ LBH.AdultSocialCare.Api/V1/Infrastructure/RequestFeatures/Extensions/GatewayQueryExtensions.cs
@@ -4,6 +4,7 @@ using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.Common;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Linq;
+using System.Runtime.InteropServices;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.CarePackages;
 
 namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Extensions
@@ -32,17 +33,28 @@ namespace LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Extensions
         public static IQueryable<User> FilterAppUsers(this IQueryable<User> users, string searchTerm = "") =>
             users.Where(u => (string.IsNullOrEmpty(searchTerm) || (EF.Functions.ILike(u.Name, $"%{searchTerm}%") || EF.Functions.ILike(u.Email, $"%{searchTerm}%"))));
 
-        public static IQueryable<CarePackage> FilterCareChargeCarePackageList(this IQueryable<CarePackage> carePackages, string status, string firstName, string lastName,
-            DateTime? dateOfBirth, string postCode, int? mosaicId, DateTimeOffset? modifiedAt, Guid? modifiedBy) =>
-            carePackages.Where(c =>
-                (firstName != null ? c.ServiceUser.FirstName.ToLower().Contains(firstName.ToLower()) : c.Equals(c)) &&
-                (lastName != null ? c.ServiceUser.LastName.ToLower().Contains(lastName.ToLower()) : c.Equals(c)) &&
-                (dateOfBirth.Equals(null) || c.ServiceUser.DateOfBirth == dateOfBirth) &&
-                (postCode != null ? c.ServiceUser.PostCode.ToLower().Contains(postCode.ToLower()) : c.Equals(c)) &&
-                (mosaicId.Equals(null) || c.ServiceUser.HackneyId == mosaicId) &&
+        public static IQueryable<CarePackage> FilterCareChargeCarePackageList(this IQueryable<CarePackage> carePackages,
+            string status, Guid? modifiedBy, string orderByDate)
+        {
+            var filteredList = carePackages.Where(c =>
                 (modifiedBy.Equals(null) || c.UpdaterId == modifiedBy) &&
-                (modifiedAt.Equals(null) || c.DateUpdated == modifiedAt) &&
-                (status != null ? c.Reclaims.Any(r => r.Type == ReclaimType.CareCharge) == (status == "Existing") : c.Equals(c)));
+                (status != null ? c.Reclaims.Any(r => r.Type == ReclaimType.CareCharge && r.SubType != ReclaimSubType.CareChargeProvisional) == (status == "Existing") : c.Equals(c)));
+
+            switch (orderByDate)
+            {
+                case "desc":
+                    filteredList = filteredList.OrderByDescending(s => s.DateCreated);
+                    break;
+                case "asc":
+                    filteredList = filteredList.OrderBy(s => s.DateCreated);
+                    break;
+                default:
+                    filteredList = filteredList.OrderBy(s => s.DateCreated);
+                    break;
+            }
+
+            return filteredList;
+        }
 
         public static IQueryable<ServiceUser> FilterServiceUser(this IQueryable<ServiceUser> serviceUsers, string firstName, string lastName,
             string postCode, DateTime? dateOfBirth, int? hackneyId) =>
