+++ LBH.AdultSocialCare.Api/V1/UseCase/CarePackages/Concrete/GetCareChargePackagesUseCase.cs
@@ -3,6 +3,7 @@ using LBH.AdultSocialCare.Api.V1.Boundary.CarePackages.Response;
 using LBH.AdultSocialCare.Api.V1.Boundary.Common.Response;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways.CarePackages.Interfaces;
+using LBH.AdultSocialCare.Api.V1.Gateways.Common.Interfaces;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.RequestFeatures.Parameters;
 using LBH.AdultSocialCare.Api.V1.UseCase.CarePackages.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Common.Interfaces;
@@ -11,16 +12,16 @@ namespace LBH.AdultSocialCare.Api.V1.UseCase.CarePackages.Concrete
 {
     public class GetCareChargePackagesUseCase : IGetCareChargePackagesUseCase
     {
-        private readonly ICarePackageReclaimGateway _carePackageReclaimGateway;
+        private readonly ICareChargesGateway _careChargesGateway;
 
-        public GetCareChargePackagesUseCase(ICarePackageReclaimGateway carePackageReclaimGateway)
+        public GetCareChargePackagesUseCase(ICareChargesGateway careChargesGateway)
         {
-            _carePackageReclaimGateway = carePackageReclaimGateway;
+            _careChargesGateway = careChargesGateway;
         }
 
         public async Task<PagedResponse<CareChargePackagesResponse>> GetCareChargePackages(CareChargePackagesParameters parameters)
         {
-            var result = await _carePackageReclaimGateway.GetCareChargePackages(parameters).ConfigureAwait(false);
+            var result = await _careChargesGateway.GetCareChargePackages(parameters).ConfigureAwait(false);
 
             return new PagedResponse<CareChargePackagesResponse>
             {
